Refactoring json presenter
